requirements.txt
discord.py
dnspython
GitPython
motor
lavalink
aiohttp
websockets
pyttanko
redis
jishaku
sentry-sdk
pyfiglet
